[chore](compile) Fix some implicit conversion issues
be/src/cloud/cloud_internal_service.cpp
be/src/cloud/pb_convert.cpp
be/src/io/file_factory.cpp
be/src/io/fs/hdfs_file_reader.cpp
be/src/io/fs/hdfs_file_writer.cpp
be/src/io/fs/local_file_writer.cpp
be/src/io/fs/s3_file_writer.cpp
be/src/io/fs/s3_file_writer.h

+++ be/src/cloud/cloud_internal_service.cpp
@@ -23,6 +23,7 @@
 #include "io/cache/block_file_cache_factory.h"
 
 namespace doris {
+#include "common/compile_check_begin.h"
 
 CloudInternalServiceImpl::CloudInternalServiceImpl(CloudStorageEngine& engine, ExecEnv* exec_env)
         : PInternalService(exec_env), _engine(engine) {}
@@ -82,7 +83,7 @@ void CloudInternalServiceImpl::get_file_cache_meta_by_tablet_id(
         auto rowsets = tablet->get_snapshot_rowset();
         std::for_each(rowsets.cbegin(), rowsets.cend(), [&](const RowsetSharedPtr& rowset) {
             std::string rowset_id = rowset->rowset_id().to_string();
-            for (int64_t segment_id = 0; segment_id < rowset->num_segments(); segment_id++) {
+            for (int32_t segment_id = 0; segment_id < rowset->num_segments(); segment_id++) {
                 std::string file_name = fmt::format("{}_{}.dat", rowset_id, segment_id);
                 auto cache_key = io::BlockFileCache::hash(file_name);
                 auto* cache = io::FileCacheFactory::instance()->get_by_path(cache_key);
@@ -104,5 +105,6 @@ void CloudInternalServiceImpl::get_file_cache_meta_by_tablet_id(
         });
     }
 }
+#include "common/compile_check_end.h"
 
 } // namespace doris
